Classic Noob Mistake
You could almost accuse me of making this mistake on purpose just so
I'd get a nice clean diff in Git commit style, showing what happens
when you forget to set the owner's type in a blank nib you created. (In
my case it's the WPTopLevelViewController_iPad.xib).  The iPhone
version was created by the new-template-project and so it was
automatically set. The iPad one was created as a new blank nib.