Sample should actually show you using the Nav controller.
This code adds a new sub-level view so that we have somewhere to
navigate to.
Then we navigate to it.  In so doing we learn that
presentViewController is not the
right message to send, it's pushViewController, which implies we're
creating
a stack. Which we are.  Aha.  NavController has a stack of presenting
controllers
and going back is popping an item off a stack.